requirements.txt
streamlit
openai
python-docx
pdfplumber
jinja2
weasyprint
